requirements.txt
langchain
langchain-openai
mcp
fastapi
uvicorn
python-dotenv
sqlalchemy
openai
langfuse
